Reorganizes OceanProblems and add InitialValueProblem
