+++ src/Ocean/Ocean.jl
@@ -22,7 +22,7 @@ include("OceanBC.jl")
 include("HydrostaticBoussinesq/HydrostaticBoussinesqModel.jl")
 include("ShallowWater/ShallowWaterModel.jl")
 include("SplitExplicit/SplitExplicitModel.jl")
-include("OceanProblems/SimpleBoxProblem.jl")
 include("SplitExplicit01/SplitExplicitModel.jl")
+include("OceanProblems/OceanProblems.jl")
 
 end
